rust: exchange mod
Most part of the code has been commented in exchange.rs because it uses
structs that will be moved or implemented in further commits
